refactor(eval): extract cli parsing helpers
